requirements.txt
streamlit==1.41.1
anthropic==0.49.0
pydantic==2.11.1
python-dotenv==1.1.0
pdfminer.six==20231228
python-docx==1.1.2
tavily-python==0.5.0
requests==2.32.3
beautifulsoup4==4.13.3
pytest==8.3.4
pytest-mock==3.14.0
